backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
librosa==0.10.1
soundfile==0.12.1
numpy==1.26.3
scipy==1.11.4
dtw-python==1.3.0
pandas==2.1.4
aiofiles==23.2.1
